Rename package to 'scribbles' and change license
This deliberately forces a change in the way the library is called, as
well as take advantage of the fact that the 'scribbles' package name is
still free on PyPi (too many instances of 'boiler' and 'scripts' and
their derivatives).

The Python version is also upped to 3.8 and ideally will stay at this
version for the forseeable future. No point wasting time doing backwards
maintenance for systems with EOL Python - at that point in time, the
system itself ought to be upgraded to mitigate old vulnerabilities.